requirements.txt
google-genai
streamlit
PyMuPDF
python-dotenv
PyYAML

pandas
